refactor socket gateway and socket service
